[takea-look#67] Circuit 초기 세팅
