+++ app/src/main/java/my/takealook/di/CircuitModule.kt
@@ -1,0 +1,33 @@
+package my.takealook.di
+
+import com.slack.circuit.foundation.Circuit
+import com.slack.circuit.runtime.presenter.Presenter
+import com.slack.circuit.runtime.ui.Ui
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import dagger.multibindings.Multibinds
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class CircuitModule {
+
+    @Multibinds
+    abstract fun presenterFactories(): Set<Presenter.Factory>
+
+    @Multibinds
+    abstract fun uiFactories(): Set<Ui.Factory>
+
+    companion object {
+        @[Provides Singleton]
+        fun provideCircuit(
+            presenterFactories: @JvmSuppressWildcards Set<Presenter.Factory>,
+            uiFactories: @JvmSuppressWildcards Set<Ui.Factory>,
+        ): Circuit = Circuit.Builder()
+            .addPresenterFactories(presenterFactories)
+            .addUiFactories(uiFactories)
+            .build()
+    }
+}
