+++ build-logic/convention/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
     compileOnly(libs.android.tools.common)
     compileOnly(libs.compose.gradlePlugin)
     compileOnly(libs.kotlin.gradlePlugin)
+    compileOnly(libs.kspplugin)
 }
 
 gradlePlugin {
