+++ build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.hilt) apply false
     alias(libs.plugins.kotlin.serialization) apply false
+    alias(libs.plugins.kotlin.parcelize) apply false
 
     alias(libs.plugins.takealook.android.application.compose) apply false
     alias(libs.plugins.takealook.android.library.compose) apply false
